added logger class in new file
