--- cli_output.txt
@@ -1,22 +1,0 @@
-
-Pheno file is not separate from sequence file, assuming seq_file is either a .csv or a .xlsx file.
-Reading .xlsx file.
-['0000000000000002333', '0000000000000002333', '0000000000000233333', '0000000000000223333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000022333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000022333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000002333', '0000000000000000233', '0000000000000000233', '0000000000000000233', '0000000000000000023', '0000000000000000223', '0000000000000000223', '0000000000000000002', '0000000000000000002', '0000000001000023333', '0100000000000023333', '0000000100000002333', '0010000000000000233', '1000000000000000233', '0000010000000000002', '0000000000010000002', '0000000001000000002', '0100100233333333333', '0010001000000002333', '0010000100000002333', '0010000000001002333', '0011000000000022333', '1000001000000000002', '0010010000001023333', '0000010000110000233', '0010010001000000002', '0110000100010233333', '0000101000000010102', '0001001010110023333', '0010010000011010233', '0000000000000102333', '0000011023333333333', '0000001100233333333', '0110010000233333333', '1000100010002233333', '0111000000000002333', '1001101102333333333']
-
-Phenotype Values (Fest):
-[5.87 3.73 5.87 3.63 4.14 3.72 5.87 5.87 5.87 4.4  3.71 3.71 3.77 5.87
- 5.87 4.02 5.87 5.87 5.87 3.73 4.07 3.72 3.51 3.76 5.87 4.04 5.87 5.87
- 0.24 5.87 5.87 3.57 3.82 4.05 5.87 4.05 5.87 5.87 3.91 5.87 3.7  3.82
- 5.87 5.   5.87 5.87 2.08 5.   3.91 4.02 4.05 4.02 3.61 5.87 3.97 4.44
- 5.87 5.87 3.73 4.86 4.76 4.02 4.44 3.54 5.87 3.94 5.87 3.6  5.87 4.04
- 3.56 3.63 5.87 5.87 5.87 5.87 2.08 5.58 5.87 3.64 5.87]
-Will be computing 81 sequences with 19 sites, and each vector will be 4-dimensional.
-
-Groupings according to --alphbt_input:
-{0: SYG | 1: R | 2: AFILM | 3: n}
-rf1d form of alphabet input:
-SYG,R,z,n
-"z" is AFILM
-Items not in sequence dataset: ['N', 'D', 'C', 'E', 'Q', 'H', 'K', 'P', 'T', 'W', 'V']
-computed mean
-computed variance
